backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-google-genai
openai
structlog
pydantic
python-multipart
rarfile
py7zr
requests
psutil
click>=8.1.0,<9.0.0
charset-normalizer>=3.0.0,<4.0.0
annotated-types>=0.7.0,<1.0.0
anyio>=4.11.0,<5.0.0
cachetools>=5.5.0,<6.0.0
certifi>=2025.8.0,<2026.0.0
colorama>=0.4.6,<1.0.0
distro>=1.9.0,<2.0.0
filetype>=1.2.0,<2.0.0
google-ai-generativelanguage>=0.7.0,<1.0.0
google-api-core>=2.25.0,<3.0.0
google-auth>=2.40.0,<3.0.0
googleapis-common-protos>=1.70.0,<2.0.0
greenlet>=3.2.0,<4.0.0
grpcio>=1.75.0,<2.0.0
grpcio-status>=1.75.0,<2.0.0
h11>=0.16.0,<1.0.0
httpcore>=1.0.0,<2.0.0
httpx>=0.28.0,<1.0.0
idna>=3.10,<4.0.0
inflate64>=1.0.3,<2.0.0
jiter>=0.11.0,<1.0.0
jsonpatch>=1.33,<2.0.0
jsonpointer>=3.0.0,<4.0.0
langchain-core>=0.3.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
langgraph>=0.6.0,<1.0.0
langgraph-checkpoint>=2.1.0,<3.0.0
langgraph-prebuilt>=0.6.0,<1.0.0
langgraph-sdk>=0.2.0,<1.0.0
langsmith>=0.4.0,<1.0.0
multivolumefile>=0.2.3,<1.0.0
orjson>=3.11.0,<4.0.0
ormsgpack>=1.10.0,<2.0.0
packaging>=25.0,<26.0
proto-plus>=1.26.0,<2.0.0
protobuf>=6.32.0,<7.0.0
pyasn1>=0.6.0,<1.0.0
pyasn1_modules>=0.4.0,<1.0.0
pybcj>=1.0.6,<2.0.0
pycryptodomex>=3.23.0,<4.0.0
pydantic_core>=2.33.0,<3.0.0
PyPDF2>=3.0.0,<4.0.0
pyppmd>=1.2.0,<2.0.0
PyYAML>=6.0.0,<7.0.0
pyzstd>=0.17.0,<1.0.0
regex>=2025.9.0,<2026.0.0
requests-toolbelt>=1.0.0,<2.0.0
rsa>=4.9.0,<5.0.0
setuptools>=78.1.0,<79.0.0
sniffio>=1.3.0,<2.0.0
SQLAlchemy>=2.0.0,<3.0.0
starlette>=0.48.0,<1.0.0
tenacity>=9.1.0,<10.0.0
texttable>=1.7.0,<2.0.0
tiktoken>=0.11.0,<1.0.0
tqdm>=4.67.0,<5.0.0
typing-inspection>=0.4.1,<1.0.0
typing_extensions>=4.15.0,<5.0.0
urllib3>=2.5.0,<3.0.0
wheel>=0.45.0,<1.0.0
xxhash>=3.5.0,<4.0.0
zstandard>=0.25.0,<1.0.0